Update to Xpdf 0.92.
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@1695 7a7537e8-13f0-0310-91df-b6672ffda945

+++ pdftops/Page.h
@@ -99,7 +99,7 @@ public:
   Object *getContents(Object *obj) { return contents.fetch(obj); }
 
   // Display a page.
-  void display(OutputDev *out, int dpi, int rotate,
+  void display(OutputDev *out, double dpi, int rotate,
 	       Links *links, Catalog *catalog);
 
 private:
